Fefactor Modal and MenuList using custom hooks
