+++ src/utils/hooks/useModal.ts
@@ -1,0 +1,9 @@
+import { useState } from 'react'
+
+const useModal = () => {
+  const [open, setOpen] = useState(false)
+
+  return { open, setOpen }
+}
+
+export default useModal
